requirements.txt
crewai
langchain
langchain-community
langchain-google-genai
langchain-text-splitters
chromadb
pypdf
litellm
sqlalchemy
python-dotenv
google-generativeai
fastapi
uvicorn